move music commands to their own file
